Add a Billing settings placeholder.
