+++ services/website/src/components/settings/website/billing.svelte
@@ -1,0 +1,19 @@
+<div class="mt-5 md:mt-0 md:col-span-2">
+  <form on:submit|preventDefault={() => {}}>
+    <div class="shadow overflow-hidden sm:rounded-md">
+      <div class="px-4 py-5 bg-white sm:p-6">
+        <fieldset>
+          <legend class="text-base leading-6 font-medium text-gray-900">Billing</legend>
+          <div class="mt-4 space-y-4">
+            Coming soon: Billing for websites with 10k+ events per month.
+          </div>
+        </fieldset>
+      </div>
+      <div class="px-4 py-3 bg-gray-50 text-right sm:px-6">
+        <button type="submit" class="py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 shadow-sm hover:bg-indigo-500 focus:outline-none focus:shadow-outline-blue focus:bg-indigo-500 active:bg-indigo-600 transition duration-150 ease-in-out">
+          Save
+        </button>
+      </div>
+    </div>
+  </form>
+</div>
